Move common fns to virtwrap/api/util.go
