+++ pkg/liveupdate/memory/memory.go
@@ -25,7 +25,6 @@ import (
 	v1 "kubevirt.io/api/core/v1"
 
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
-	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/vcpu"
 
 	"k8s.io/apimachinery/pkg/api/resource"
 )
@@ -112,14 +111,14 @@ func BuildMemoryDevice(vmi *v1.VirtualMachineInstance) (*api.MemoryDevice, error
 
 	pluggableMemory := domain.Memory.MaxGuest.DeepCopy()
 	pluggableMemory.Sub(*vmi.Status.Memory.GuestAtBoot)
-	pluggableMemorySize, err := vcpu.QuantityToByte(pluggableMemory)
+	pluggableMemorySize, err := api.QuantityToByte(pluggableMemory)
 	if err != nil {
 		return nil, err
 	}
 
 	requestedHotPlugMemory := domain.Memory.Guest.DeepCopy()
 	requestedHotPlugMemory.Sub(*vmi.Status.Memory.GuestAtBoot)
-	pluggableMemoryRequested, err := vcpu.QuantityToByte(requestedHotPlugMemory)
+	pluggableMemoryRequested, err := api.QuantityToByte(requestedHotPlugMemory)
 	if err != nil {
 		return nil, err
 	}
