+++ pkg/virt-launcher/virtwrap/converter/network.go
@@ -23,12 +23,10 @@ import (
 	"fmt"
 
 	v1 "kubevirt.io/api/core/v1"
-	"kubevirt.io/client-go/log"
 
 	netvmispec "kubevirt.io/kubevirt/pkg/network/vmispec"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/arch"
-	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/vcpu"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/device"
 )
 
@@ -60,7 +58,7 @@ func CreateDomainInterfaces(vmi *v1.VirtualMachineInstance, c *ConverterContext)
 			Alias: api.NewUserDefinedAlias(iface.Name),
 		}
 
-		if queueCount := uint(CalculateNetworkQueues(vmi, ifaceType)); queueCount != 0 {
+		if queueCount := uint(api.CalculateNetworkQueues(vmi, ifaceType)); queueCount != 0 {
 			domainIface.Driver = &api.InterfaceDriver{Name: "vhost", Queues: &queueCount}
 		}
 
@@ -123,32 +121,6 @@ func indexNetworksByName(networks []v1.Network) map[string]*v1.Network {
 		netsByName[network.Name] = network.DeepCopy()
 	}
 	return netsByName
-}
-
-func CalculateNetworkQueues(vmi *v1.VirtualMachineInstance, ifaceType string) uint32 {
-	if ifaceType != v1.VirtIO {
-		return 0
-	}
-	return NetworkQueuesCapacity(vmi)
-}
-
-func NetworkQueuesCapacity(vmi *v1.VirtualMachineInstance) uint32 {
-	if !isTrue(vmi.Spec.Domain.Devices.NetworkInterfaceMultiQueue) {
-		return 0
-	}
-
-	cpuTopology := vcpu.GetCPUTopology(vmi)
-	queueNumber := vcpu.CalculateRequestedVCPUs(cpuTopology)
-
-	if queueNumber > multiQueueMaxQueues {
-		log.Log.V(3).Infof("Capped the number of queues to be the current maximum of tap device queues: %d", multiQueueMaxQueues)
-		queueNumber = multiQueueMaxQueues
-	}
-	return queueNumber
-}
-
-func isTrue(networkInterfaceMultiQueue *bool) bool {
-	return (networkInterfaceMultiQueue != nil) && (*networkInterfaceMultiQueue)
 }
 
 func translateModel(useVirtioTransitional *bool, bus string, archString string) string {
