+++ pkg/virt-launcher/virtwrap/converter/vcpu/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
     deps = [
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
         "//pkg/util/hardware:go_default_library",
+        "//pkg/virt-launcher/util:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
